map JS and CSS
