+++ app/assets/stylesheets/components/_map.scss
@@ -2,7 +2,8 @@
 #map {
   position: sticky;
   top: 0;
-  border: 4px $light-blue solid;
+  border: 2px $blue solid;
+ 
 }
 
 .mapboxgl-popup {
