+++ app/assets/stylesheets/components/_navbar.scss
@@ -1,4 +1,5 @@
 .navbar-lewagon {
+  background-color: $blue;
   justify-content: space-between;
 }
 
@@ -14,16 +15,14 @@
   border-radius: 50%;
 }
 
-
 .nav-link{
   font-family: $body-font;
-  color: $dark-blue;
-  font-weight: bolder;
+  color: $beige;
   font-size: 18px;
-}
-
-.nav-link:hover{
-  color: $blue;
+  padding-left: 5px;
+  .nav-link:hover{
+    color: $dark-blue;
+  }
 }
 
 .nav-item{
