+++ app/assets/stylesheets/components/_runner_card.scss
@@ -29,22 +29,26 @@
 }
 
 .popup-card{
-  display: none;
+  background: $white;
+  border: 3px solid $blue;
+  border-radius: 24px;
+  // background-image: url('card-bg.jpg');
+  display: block;
   position: fixed;
   z-index: 1;
   left: 0;
   top: 0;
-  width: 100px;
-  height: 100px;
+  width: 400px;
+  height: 650px;
   overflow: auto;
-  padding-top: 16px;
-  background-color: $light-blue;
+  // background-color: $light-blue;
   margin: 8px 16px;
+  padding: 40px 32px;
   text-align: center;
 }
 
 .close {
-  color: #aaa;
+  color: $blue;
   float: right;
   font-size: 28px;
   font-weight: bold;
@@ -52,14 +56,17 @@
 
 .close:hover,
 .close:focus {
-  color: black;
+  color: grey;
   text-decoration: none;
   cursor: pointer;
 }
 
 .runner-card-img{
-  height:100px;
-  width:100px;
+  height:200px;
+  width:200px;
+  object-fit: cover;
+  background: cover;
+  border-radius: 16px;
   margin-bottom:20px;
   margin-left: 5px;
 }
