+++ app/assets/stylesheets/pages/_home.scss
@@ -65,4 +65,3 @@
   font-size: 20px;
   padding-top: 40px;
 }
-
