+++ app/assets/stylesheets/pages/_runners.scss
@@ -3,19 +3,22 @@
   width: 100%;
 }
 
+.text-header{
+  text-align: center;
+}
+
 .container{
   width: 100vw;
   position: relative;
- height: 100px;
- margin-top: 100px;
+  height: 100px;
+  margin-top: 80px;
 }
 
 .header-style{
-  color: black;
-  font-size: 24px;
-  font-weight: bold;
+  color: $dark-blue;
+  font-size: 30px;
   padding-bottom: 20px;
-  font-family: $body-font;
+  font-family: $logo-font;
 }
 
 .runners-list{
