+++ app/javascript/plugins/initcard.js
@@ -1,0 +1,29 @@
+const init_popup = () => { 
+  // Get the modal
+  const modal = document.getElementById("popup-card");
+
+  // Get the button that opens the modal
+  const btn = document.getElementById("runner-click");
+
+  // Get the <span> element that closes the modal
+  const close = document.getElementsByClassName("close")[0];
+
+  // When the user clicks on the button, open the modal
+  btn.onclick = function() {
+    modal.style.display = "block";
+  };
+
+  // When the user clicks on <span> (x), close the modal
+  close.onclick = function() {
+    modal.style.display = "none";
+  };
+
+  // When the user clicks anywhere outside of the modal, close it
+  window.onclick = function(event) {
+    if (event.target == modal) {
+      modal.style.display = "none";
+    };
+  };
+};
+
+export { init_popup };
