--- app/javascript/plugins/runner_card.js
@@ -1,28 +1,0 @@
-const init_popup = () => { 
-// Get the modal
-var modal = document.getElementById("popup-card");
-
-// Get the button that opens the modal
-var btn = document.getElementById("runner-click");
-
-// Get the <span> element that closes the modal
-var span = document.getElementsByClassName("close")[0];
-
-// When the user clicks on the button, open the modal
-btn.onclick = function() {
-  modal.style.display = "block";
-}
-
-// When the user clicks on <span> (x), close the modal
-span.onclick = function() {
-  modal.style.display = "none";
-}
-
-// When the user clicks anywhere outside of the modal, close it
-window.onclick = function(event) {
-  if (event.target == modal) {
-    modal.style.display = "none";
-  }
-}
-
-// export { init_popup };
