+++ app/views/runners/index.html.erb
@@ -1,15 +1,8 @@
 <div class="map-page">
-
-  <div class="container card-container">
-    <h1 class="header-style">Runners Near Me</h1>
-    <div class="search-option">
-       <select class="distance select2" name="distance">
-        <option value="">within 5km</option>
-        <option value="">within 10km</option>
-        <option value="">within 25km</option>
-        <option value="">within 50km</option>
-      </select>
-
+<div class="container card-container">
+  <div class="text-header">
+   <h1 class="header-style">Find out other Runners near you</h1>
+   <div class="search-option">
       <select class="running-level select2" name="running-level">
         <option value="">Fast</option>
         <option value="">Averge</option>
@@ -31,12 +24,12 @@
         <option value="">Afternoon</option>
         <option value="">Evening</option>
       </select>
-
       <%= link_to "Find" %>
     </div>
-
+    </div>
+     
     <div class="row">
-    <div class="col-md-7 map">
+      <div class="col-7 map">
         <div id="map"
               style="width: 100%; height: 100vh;"
               data-markers="<%= @markers.to_json %>"
@@ -44,24 +37,34 @@
         </div>
       </div>
 
-      <div class="runner-cards col-md-5">
-      
+     <div class="col-4 runner-cards">
         <% @runners.each do |runner| %>
           <% if runner == current_user %>
           <% else %>
            <div class="runner-card" id="runner-click" style="background-image: linear-gradient(rgba(0,0,0,0.3), rgba(0,0,0,0.3)), url('<%=Cloudinary::Utils.cloudinary_url(runner.image.key)%>'); "data-marker-id="<%= runner.id %>">
             <strong><%= runner.first_name%> <%= runner.last_name%></strong>
             </div>
-  
+            
+          <div class="popup-card" id="popup-card">
+           <span class="close">x</span>
+              <%= cl_image_tag runner.image.key, crop: :fill, class: "runner-card-img" %>
+              <div class="runner-card-body">
+              <strong><%= runner.first_name%> <%= runner.last_name%></strong><br/>
+              <%= runner.gender %><br/>
+              <%= runner.address %><br/><br/>
+              <%= runner.introduction%><br/><br/>
+              <%=link_to "Buddy request",
+                runner_requests_path(runner),
+                method: :post,
+                data: { confirm: "Do you want to send a buddy request?", commit: 'Sure!' },
+                class: "btn btn-md btn-primary buddy-request-button"
+              %>
+              </div>
+        </div>
           <% end %>
         <% end %>
-      </div>
+     </div>
 
- 
-
-
-    </div>
-
-    </div>
   </div>
+</div>
 </div>
